requirements.txt
# Core dependencies
pandas>=2.2.0
numpy>=1.24.0
polars>=0.20.0  # Optional, for large transforms
httpx>=0.24.0
tenacity>=8.2.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-httpx>=0.24.0

# Code quality
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Research
statsmodels>=0.14.0
scipy>=1.11.0
